Fixed the various makefile.win so that building in debug (sc. nmake /f makefile.win CFG=debug) now works.  Lots of warnings about STL truncation and all that stuff, but it works.
Fixed the GeomRenderer object; it now seems to work perfectly!  Commented
out code in ttexgen exercises various paths in the GeomRenderer, and they
all succeeded.

My next step is to add to geomutil an object to generate a sphere and use
that with GeomRenderer in ttexgen instead of glutSolidSphere, which should
also allow me to iterate over the various paths and make ttexgen more
rigorous (right now it only tests, I assume, the glVertex path, since that's
probably what glutSolidSphere uses.)