Fix some issues in the fbo test.
Convert the big CHECK_FRAMEBUFFER_STATUS macro into a function.
Check if the framebuffer is incomplete (GL_FRAMEBUFFER_UNSUPPORTED_EXT)
before rendering/testing and skip the test if so.  Drivers may not support
all variations of depth/stencil attachments and will set the FB status to
UNSUPPORTED.  For now, we just skip the test when that happens.  Ideally we
should try to setup a different depth/stencil combination and retry.
Someday...