dsconfig: try exact config match in match() before searching
This fixes a failed assertion in readpixPerf when comparing results.

readpixPerf produces different/extra test results if the visual config
has Z and/or stencil.  The comparison code was trying to compare results
from different visual configs and finding a different number of results
so an assertion failed.

Now we first look for an exact visual config match before searching for
a closest match.